Hackers_of_FII/Satoshi_Nakamoto/pyproject.toml
[tool.poetry]
name = "repl_python3_SatoshiNakamotov1"
version = "0.1.0"
description = "Custom made bot for Hackers of FII"
authors = ["Ionut Rosca <ionut.rosca@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.19.2"
Flask = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
